backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
bcrypt==4.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
aiosqlite==0.19.0
httpx==0.25.2
pyjwt==2.8.0
cryptography==41.0.7
pyyaml==6.0.1
markdown==3.5.2
